Fix/running fix (#7)
* feat(auth): implement writeAccessToken method in AuthLocalDataSource and update RefreshTokenInterceptor class

* feat(tiers): update tier display logic and remove unused imports

* feat(ranks): remove unused rank card images and refactor rank handling logic

* fix(achievement_content): update text styles and adjust accent font weight
fix(main): reduce max toast limit from 3 to 1

* feat(exercise): refactor exercise handling and add type support for program exercises

* Refactor running service client and tracking engine for improved async handling

- Updated `MockPedometerTrackingEngine` and `PedometerTrackingEngine` to use async `stop()` and `dispose()` methods for better resource management.
- Introduced `RunningServiceProtocol` for strict message parsing between the service and client, ensuring type safety and error handling.
- Enhanced `RunningServiceClient` to handle warm-up and session management more robustly, including improved error handling and logging.
- Added tests for `RunningTrackerCubit` and `RunningServiceClient` to ensure correct behavior during session management and error scenarios.
- Refactored `RunningSessionManager` to handle session state more effectively, ensuring proper cleanup and resource release.
- Updated exception handling in `running_service_exceptions.dart` for clearer error reporting.

* Refactor workout congratulations feature

- Removed WorkoutCongratulationsCubit and associated state management.
- Deleted WorkoutAchievementPage and WorkoutSummaryPage, consolidating their logic into WorkoutCongratulationsPage.
- Introduced WorkoutCongratulationsPage to handle both achievement and summary displays.
- Updated analytics logging for milestone achievements and summary completion.
- Adjusted UI components to reflect new structure and ensure proper state handling.
- Added comprehensive tests for the new WorkoutCongratulationsPage functionality.

* feat(running): refactor workout session handling to support program exercises

* feat(workout): update workout titles and enhance layout for better responsiveness

* Add tests for achievements, home, leaderboard, and subscription features

- Implement tests for AchievementsCubit to track faction changes and handle user logout.
- Enhance HomeCubit tests to include user changes and ensure rank preservation.
- Update FactionsLeaderboardCubit tests to verify faction changes and user logout behavior.
- Add ImmortalForgesCubit tests for user faction switching and repository interactions.
- Extend UsersLeaderboardCubit tests to check for public user changes and logout handling.
- Introduce email content tests to validate email submission behavior and error handling.
- Modify SubscriptionCubit tests to react only to identity and subscription changes.

* feat: Enhance Pedometer and GPS Tracking Engines with Robust Error Handling

- Introduced a new `RunningSensorAvailability` service to check for sensor availability and permissions.
- Updated `PedometerTrackingEngine` to handle terminal failures and health checks more effectively.
- Refactored error handling in `RunningSessionManager` to finalize laps before forwarding terminal errors.
- Added tests for terminal engine failures and ensured proper session management during errors.
- Enhanced exception handling with new `TrackingEngineFailureException` for better clarity on failure reasons.
- Improved the handling of recoverable errors in both Pedometer and GPS tracking engines.

* feat(running): implement session status management and error handling in tracker

* feat: update body weight handling to use double type and improve formatting

* chore: increment build number to 10 (0.30.0)

+++ lib/features/quiz/data/models/quiz_answers.dart
@@ -14,7 +14,7 @@ sealed class QuizAnswers with _$QuizAnswers {
 
     required MeasurementSystem measurementSystem,
 
-    required int bodyWeight,
+    required double bodyWeight,
 
     required MainGoal mainGoal,
